Tidy up the code
